테스트 오류로 ImageFileService delete 부분을 save 참고하여 조금 수정하고 사용했습니다..!

+++ src/main/java/homes/has/service/ImageFileService.java
@@ -130,6 +130,14 @@ public class ImageFileService {
     }
 
     public void delete(ImageFile imageFile){
+        if (imageFile == null) {
+            return;
+        }
+        String path = imageFile.getPath();
+        File file = new File(path);
+        if (file.exists()) {
+            file.delete();
+        }
         imageFileRepository.delete(imageFile);
     }
